Refactor endpoints
